+++ Test/Splices.hs
@@ -227,4 +227,28 @@ imp_inst_test4 = [d| type instance Dec8 Int = Bool |]
 type family Poly (a :: k) :: *
 type instance Poly x = Int
 
+flatten_dvald_test = [| let (a,b,c) = ("foo", 4, False) in
+                        show a ++ show b ++ show c |]
+
+rec_sel_test = [d| data RecordSel a = forall b. (Show a, Eq b) =>
+                                      MkRecord { recsel1 :: (Int, a)
+                                            , recsel_naughty :: (a, b)
+                                            , recsel2 :: (forall b. b -> a)
+                                            , recsel3 :: Bool }
+                                    | MkRecord2 { recsel4 :: (a, a) } |]
+rec_sel_test_num_sels = 4 :: Int   -- exclude naughty one
+
+testRecSelTypes :: Int -> Q Exp
+testRecSelTypes n = do
+  VarI _ ty1 _ _ <- reify (mkName ("Test.DsDec.recsel" ++ show n))
+  VarI _ ty2 _ _ <- reify (mkName ("Test.Dec.recsel"   ++ show n))
+  let ty1' = return $ unqualify ty1
+      ty2' = return $ unqualify ty2
+  [| let x :: $ty1'
+         x = undefined
+         y :: $ty2'
+         y = undefined
+     in
+     $(return $ VarE $ mkName "hasSameType") x y |]
+